Move UserFilterSet to new file to seperate it
